+++ Source/Frontend/UI/Components/Engine Config/EngineControls/BlastGeneratorEngineControl.resx
