+++ Source/Frontend/UI/Components/Engine Config/EngineControls/DistortionEngineControl.resx
