+++ Source/Frontend/UI/Components/Engine Config/EngineControls/FreezeEngineControl.resx
