+++ Source/Frontend/UI/Components/Engine Config/EngineControls/NightmareEngineControl.cs
@@ -1,4 +1,4 @@
-namespace RTCV.UI.Components.EngineConfig.Engines
+namespace RTCV.UI.Components.EngineConfig.EngineControls
 {
     using System;
     using System.Collections.Generic;
