+++ Source/Frontend/UI/Components/Engine Config/EngineControls/NightmareEngineControl.resx
