+++ Source/Frontend/UI/Components/Engine Config/EngineControls/VectorEngineControl.cs
@@ -1,4 +1,4 @@
-namespace RTCV.UI.Components.EngineConfig.Engines
+namespace RTCV.UI.Components.EngineConfig.EngineControls
 {
     internal partial class VectorEngine : EngineConfigControl
     {
