+++ Source/Frontend/UI/Components/Engine Config/EngineControls/VectorEngineControl.resx
